requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# WMSDatamining/cluster.py: 24
# WMSDatamining/project.py: 1
Flask == 1.1.2

# WMSDatamining/project.py: 13
Werkzeug == 1.0.1

# WMSDatamining/project.py: 6
apyori == 1.1.2

# WMSDatamining/cluster.py: 3,26,27
kmodes == 0.10.2

# WMSDatamining/cluster.py: 22
numpy == 1.18.4

# WMSDatamining/cluster.py: 21
# WMSDatamining/project.py: 4
pandas == 1.0.3

# WMSDatamining/cluster.py: 19,25
# WMSDatamining/project.py: 9,10
plotly == 4.7.1

# WMSDatamining/cluster.py: 6,7,8,9,10,11,12,13,14,15,16,17
pyclustering == 0.9.3.1

# WMSDatamining/project.py: 2
python_dotenv == 0.13.0

# WMSDatamining/cluster.py: 2,18
sklearn == 0.0
